% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate.R
\name{extrapolate_effects}
\alias{extrapolate_effects}
\alias{pull_effects}
\alias{augment_future}
\title{Tools for extrapolating effects from models}
\usage{
extrapolate_effects(mod, col_name, ...)

pull_effects(mod, col_name, ...)

augment_future(mod, newdata, col_name, ...)
}
\arguments{
\item{mod}{a model that with a \code{\link[broom]{tidy}} method}

\item{col_name}{name to match the
name of the effects}

\item{...}{currently ignored}

\item{newdata}{passed to \code{\link{predict}}}
}
\value{
an \code{\link{lm}} object
}
\description{
This function will extrapolate fixed effects for the future,
 assuming linear growth.
}
