% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{forbes}
\alias{forbes}
\title{Franchise values}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 676 rows and 8 columns.}
\usage{
forbes
}
\description{
Franchise values
}
\keyword{datasets}
