% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lkup_players}
\alias{lkup_players}
\title{Cross-reference Player IDs across systems}
\format{A data frame with one row for each player and 15 columns:
\describe{
   \item{name_last}{the person's last name}
   \item{name_first}{the person's first name}
   \item{name_given}{the person's given name}
   \item{mlbam_id}{the person's MLBAM id}
   \item{lahman_id}{the person's Lahman id}
   \item{retro_id}{the person's Retrosheet id}
   \item{bp_id}{the person's Baseball Prospectus id}
   \item{davenport_id}{the person's Davenport id}
   \item{fg_id}{the person's FanGraphs id}
   \item{cbs_id}{the person's CBS id}
   \item{espn_id}{the person's ESPN id}
   \item{nfbc_id}{the person's NFBC id}
   \item{yahoo_id}{the person's Yahoo id}
   \item{ottoneu_id}{the person's Ottoneu id}
   \item{rotowire_id}{the person's Rotowire id}
}}
\source{
\url{http://chadwick-bureau.com/the-register/}

\url{http://www.baseballprospectus.com/sortable/playerids/playerid_list.csv}

\url{http://crunchtimebaseball.com/master.csv}
}
\usage{
lkup_players
}
\description{
A data frame containing player ID mappings
}
\seealso{
\code{\link[Lahman]{Master}}
}
\keyword{datasets}
