% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lkup_teams}
\alias{lkup_teams}
\alias{standardize_team_name}
\alias{standardize_team_ids}
\title{Utilities for dealing with MLB Teams}
\usage{
lkup_teams()

standardize_team_name(x, ...)

standardize_team_ids(x)
}
\arguments{
\item{x}{character vector of team names}

\item{...}{currently ignored}
}
\value{
a \code{\link[tibble]{tibble}} with one row for each team
}
\description{
Utilities for dealing with MLB Teams
}
