% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lm_split}
\alias{lm_split}
\title{Running a linear model on data frame splits}
\usage{
lm_split(data, form, i, var_name, equal_width = TRUE, ...)
}
\arguments{
\item{data}{Dataset for modeling}

\item{form}{Formula to be passed to the linear model}

\item{i}{Number of cuts to split the data into}

\item{var_name}{Variable to base the cuts on}

\item{equal_width}{If TRUE, cuts the data based on even widths of value; if FALSE, each cut will have an approximately equal number of observations}

\item{...}{Passes lm() parameters to allow user to handle NA values}
}
\description{
Running a linear model on data frame splits
}
\examples{
lm_split(comps_hypercube, "rWAR ~ cum_rwar",i = 4, "cum_rwar", na.omit = TRUE)
}
