% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{mlb_pal}
\alias{mlb_pal}
\alias{scale_fill_mlb}
\alias{scale_color_mlb}
\title{MLB Team colors}
\usage{
mlb_pal(which = 1, names = "teamID")

scale_fill_mlb(which = 1, names = "teamID", ...)

scale_color_mlb(which = 1, names = "teamID", ...)
}
\arguments{
\item{which}{Integer giving the set of colors you want. Default is 1 for
"primary".}

\item{names}{The quoted name of the column of the tibble returned by
\code{\link{lkup_teams}} that you want to use as the names of the colors.
This will need to match the data in the variable you want to map to color.}

\item{...}{Arguments passed on to \code{ggplot2::scale_fill_manual}
\describe{
  \item{values}{a set of aesthetic values to map data values to. If this
is a named vector, then the values will be matched based on the names.
If unnamed, values will be matched in order (usually alphabetical) with
the limits of the scale. Any data values that don't match will be
given \code{na.value}.}
  \item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the \code{colour} and \code{fill} aesthetics at the
same time, via \code{aesthetics = c("colour", "fill")}.}
}}
}
\description{
MLB Team colors
}
\examples{
mlb_pal() != mlb_pal(2)

mlb_pal(names = "lahman_name")
mlb_pal(names = "teamcolors_name")
mlb_pal(2, names = "city")
}
\seealso{
\code{\link[teamcolors]{team_pal}},
\code{\link[teamcolors]{scale_fill_teams}},
\code{\link[teamcolors]{scale_color_teams}}
}
