% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mls}
\alias{mls}
\title{Major League Service Time}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1346 rows and 5 columns.}
\source{
\url{https://legacy.baseballprospectus.com/compensation/cots/}
}
\usage{
mls
}
\description{
Major League Service Time
}
\keyword{datasets}
