% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate.R
\name{mod_effect}
\alias{mod_effect}
\alias{filter_effect}
\alias{augment_future}
\title{Tools for extrapolating effects from models}
\usage{
mod_effect(mod, col_name, ...)

filter_effect(mod, col_name, ...)

augment_future(mod, newdata, col_name, ...)
}
\arguments{
\item{mod}{a model that has a \code{\link[broom]{tidy}} method}

\item{col_name}{name to match the name of the effect}

\item{...}{currently ignored}

\item{newdata}{passed to \code{\link{predict}}}
}
\value{
an \code{\link{lm}} object
}
\description{
This function will extrapolate fixed effects for the future,
 assuming linear growth.
}
