% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/war.R
\name{model_frame}
\alias{model_frame}
\title{Build a data frame suitable for projecting WAR}
\usage{
model_frame()
}
\description{
Build a data frame suitable for projecting WAR
}
\examples{
\dontrun{
test <- model_frame()
}
}
