% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msa}
\alias{msa}
\title{MSA statistics}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1839 rows and 6 columns.}
\usage{
msa
}
\description{
MSA statistics
}
\keyword{datasets}
