% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbref.R
\name{read_rwar}
\alias{read_rwar}
\title{read_rwar}
\usage{
read_rwar()
}
\value{
a data.frame consisting of rWAR
}
\description{
Retrieves rWAR from Baseball-Reference.com
}
\details{
Retrieves daily rWAR figures from Baseball-Reference.com and stores it as a data.frame
}
\examples{

# Get data from yesterday
\dontrun{
rWAR <- read_rwar()
hist(rWAR$rWAR)

# Leaders since 1954
if (require(dplyr)) {
  modern <- filter(rWAR, yearId >= 1954)
  head(arrange(modern, desc(rWAR)), 20)

  # Relationship between batting and fielding
  plot(modern$rRAA_bat, modern$rRAA_field, cex=0.5)
}
}

}
