% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/war.R
\name{war_frame}
\alias{war_frame}
\title{Build a data frame suitable for projecting WAR}
\usage{
war_frame(year = 2017, ...)
}
\description{
Build a data frame suitable for projecting WAR
}
\examples{
\dontrun{
test <- war_frame(2016)
}
}
