% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comps.R
\name{comps}
\alias{comps}
\alias{comps_universe}
\alias{comps_mod_rwar}
\alias{comps_mod_salary}
\alias{plot.comps}
\alias{plot_comps_universe}
\title{Modeling future performance}
\usage{
comps(lahman_id, horizon = as.numeric(format(Sys.Date(), "\%Y")),
  max_dist = 500, max_num = 10000)

comps_universe(lahman_id)

comps_mod_rwar(.data)

comps_mod_salary(.data)

\method{plot}{comps}(x, horizon_years = 5, ...)

plot_comps_universe(x, ...)
}
\arguments{
\item{lahman_id}{Lahman ID of the player}

\item{horizon}{Numeric description of the year from which you want to project}

\item{max_dist}{Maximum distance allowed for nearest neighbors}

\item{max_num}{Maximum number of neighbors returned}

\item{.data}{Passed to the data argument of \code{\link[stats]{lm}}}

\item{x}{A \code{comps} object}

\item{horizon_years}{How many years in advance do you want to forecast?}

\item{...}{currently ignored}
}
\description{
Modeling future performance
}
\examples{
comps("kiescbr01")
comps("kiescbr01", 2000)
comps("piazzmi01")
comps("piazzmi01", 1995)
comps("wrighda03")
comps("wrighda03", 2009)
comps_universe(c("piazzmi01", "wrighda03", "mcgwima01"))
plot(comps("wrighda03", 2009))
plot_comps_universe(comps("wrighda03", 2009))
}
