% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comps_hypercube}
\alias{comps_hypercube}
\title{Comparable Players}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 97992 rows and 62 columns.}
\usage{
comps_hypercube
}
\description{
Comparable Players
}
\keyword{datasets}
