% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contracts}
\alias{contracts}
\title{Contracts}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 350 rows and 11 columns.}
\source{
\url{https://www.spotrac.com/mlb/contracts/}
}
\usage{
contracts
}
\description{
Contracts
}
\keyword{datasets}
