% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegas.R
\name{futures}
\alias{futures}
\alias{current_probs}
\alias{glm_ws}
\alias{predict_wins}
\alias{read_ws_futures_oddsshark}
\alias{read_ws_futures_actionnetwork}
\alias{read_ws_futures_sportsline}
\title{Fetch current Vegas odds}
\usage{
current_probs()

glm_ws(from = 1998, ...)

predict_wins(mod, probs, ...)

read_ws_futures_oddsshark()

read_ws_futures_actionnetwork()

read_ws_futures_sportsline()
}
\description{
Fetch current Vegas odds
}
