% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegas.R
\name{futures}
\alias{futures}
\alias{read_ws_probs}
\alias{read_ws_futures_vegasinsider}
\alias{glm_ws}
\alias{predict_wins}
\alias{read_ws_futures_oddsshark}
\alias{read_ws_futures_actionnetwork}
\alias{read_ws_futures_sportsline}
\alias{read_ws_futures_sportsoddshistory}
\title{Fetch current Vegas futures and collect them}
\source{
\url{http://www.vegasinsider.com/mlb/odds/futures/}

\url{https://www.oddsshark.com/mlb/odds/futures}

\url{https://www.actionnetwork.com/mlb/future}

\url{https://www.sportsline.com/mlb/futures/}

\url{https://www.sportsoddshistory.com/mlb-odds/live-world-series-odds/}
}
\usage{
read_ws_probs()

read_ws_futures_vegasinsider()

glm_ws(from = 1998, ...)

predict_wins(mod, probs, ...)

read_ws_futures_oddsshark()

read_ws_futures_actionnetwork()

read_ws_futures_sportsline()

read_ws_futures_sportsoddshistory()
}
\arguments{
\item{from}{Year from which to draw historical data}

\item{...}{currently ignored}

\item{mod}{a \code{\link[stats]{glm}} object}

\item{probs}{a vector of probabilities}
}
\description{
Fetch current Vegas futures and collect them
}
\examples{
\dontrun{
read_ws_probs()
}
}
